Move SecurityFilterUtils to generic filter package
