+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/restapi/filter/SecurityFilterUtils.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.controller.athenz.filter;
+package com.yahoo.vespa.hosted.controller.restapi.filter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -12,12 +12,12 @@ import com.yahoo.jdisc.handler.ResponseHandler;
 /**
  * @author bjorncs
  */
-class SecurityFilterUtils {
+public class SecurityFilterUtils {
     private static final ObjectMapper mapper = new ObjectMapper();
 
     private SecurityFilterUtils() {}
 
-    static void sendErrorResponse(ResponseHandler responseHandler, int statusCode, String message) {
+    public static void sendErrorResponse(ResponseHandler responseHandler, int statusCode, String message) {
         Response response = new Response(statusCode);
         response.headers().put("Content-Type", "application/json");
         ObjectNode errorMessage = mapper.createObjectNode();
